setup/requirements.txt

azure-functions
azure-ai-language-conversations